+++ kong/cmd/utils/migrations.lua
@@ -1,4 +1,5 @@
 local log = require "kong.cmd.utils.log"
+local validate_path = require "kong.db.migrations.validate_path_280_300"
 
 
 local MIGRATIONS_MUTEX_KEY = "migrations"
@@ -73,6 +74,27 @@ local function up(schema_state, db, opts)
     if not opts.force and schema_state.pending_migrations then
       error("Database has pending migrations; run 'kong migrations finish'")
     end
+
+    -- Temporary: check route paths to be compatiable with 3.0 router.
+    -- This should be removed in 3.2 release once upgrading from 2.8 is no
+    -- longer supported
+    if opts.conf and opts.conf.router_flavor == "traditional_compatible" then
+      local new_migrations = schema_state.new_migrations or {}
+
+      for _, t in ipairs(new_migrations) do
+        for _, mig in ipairs(t.migrations) do
+          if mig.name == "016_280_to_300" then
+            local ok, err = validate_path[db.strategy](db.connector)
+            if not ok then
+              error(err)
+            end
+
+            break
+          end
+        end
+      end
+    end
+    -- end of temporary check
 
     if opts.force and schema_state.executed_migrations then
       log.debug("forcing re-execution of these migrations:\n%s",